Refactor MongoDatabaseProvider into new file
